*refactoring of ConvertNonZeroNode.
Move ConvertNonZeroNode from OnnxRuntime.h into ConvertNonZeroNode.cpp.
