+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -198,6 +198,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertNonMaxSuppressionNode.cpp">
       <Filter>OnnxRuntime\N</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertNonZeroNode.cpp">
+      <Filter>OnnxRuntime\N</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertConstantOfShapeNode.cpp">
       <Filter>OnnxRuntime\C</Filter>
     </ClCompile>
